move chain parsing to its own file
